requirements.txt
Django>=5.0,<5.1
python-decouple
gunicorn
psycopg[binary]
dj-database-url
requests
whitenoise
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe